Рефакторит компонент CardRating.
Выносит основную логику в компонент Rating.
CardRating теперь hoc
